refactor StreamCreate and add StreamForm component
